Move to fast_io_dsal/impl/misc/
