+++ include/fast_io_core.h
@@ -19,7 +19,7 @@
 
 #include "fast_io_concept.h"
 
-#include "fast_io_dsal/impl/push_warnings.h"
+#include "fast_io_dsal/impl/misc/push_warnings.h"
 
 #include "fast_io_core_impl/empty.h"
 #if defined(_MSC_VER) && !defined(__clang__)
@@ -95,6 +95,6 @@
 #include "fast_io_core_impl/http_header.h"
 #include "fast_io_core_impl/io_lockable.h"
 
-#include "fast_io_dsal/impl/pop_warnings.h"
+#include "fast_io_dsal/impl/misc/pop_warnings.h"
 
 #endif
